chore: clean up exports in index.ts based on feedback
